Backend/orchestration/unified_orchestration_system/requirements.txt
# Unified Orchestration System Requirements

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangChain and AI/ML
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0

# Vector database
faiss-cpu==1.7.4

# Google Gemini API
google-generativeai==0.3.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Time series forecasting
prophet>=1.1.4
statsmodels>=0.14.0
scikit-learn>=1.3.0

# PDF processing
pypdf==3.17.4
PyPDF2==3.0.1

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional utilities
pathlib2==2.3.7
typing-extensions==4.8.0
